So by this commit I have made a lot of progress. I have made a cake object that is not a power up yet but should be the basis for a dash power up. I made a smooth camera which doesn't have lag making the respawn and movement less weird. I also created a spawning system where I can respawn at different locations.
Awww yaaaaaaaaa! Still need to link the cake to the power up so it disappears and makes the script turn on. Make a platform movement script so that I can make box crushers and trickier jumps and I also need to make the level Larger. Plus I want to make elevators to show difficulty progress.